PYIC-3512 Update to only pass socialSecurityRecord if non-empty
The CRIs don't like us passing a shared claim they don't recognise and reject the authorize request.
This ensures we only pass `socialSecurityRecord` if it's non-empty
