+++ lib/src/main/java/uk/gov/di/ipv/core/library/domain/SharedClaimsResponseDto.java
@@ -12,17 +12,11 @@ public class SharedClaimsResponseDto {
     private final Set<Name> name;
     private final Set<BirthDate> birthDate;
     private final Set<Address> address;
-    private final Set<SocialSecurityRecord> socialSecurityRecord;
 
-    public SharedClaimsResponseDto(
-            Set<Name> name,
-            Set<BirthDate> birthDate,
-            Set<Address> address,
-            Set<SocialSecurityRecord> socialSecurityRecord) {
+    public SharedClaimsResponseDto(Set<Name> name, Set<BirthDate> birthDate, Set<Address> address) {
         this.name = name;
         this.birthDate = birthDate;
         this.address = address;
-        this.socialSecurityRecord = socialSecurityRecord;
     }
 
     public Set<Name> getName() {
@@ -35,9 +29,5 @@ public class SharedClaimsResponseDto {
 
     public Set<Address> getAddress() {
         return address;
-    }
-
-    public Set<SocialSecurityRecord> getSocialSecurityRecord() {
-        return socialSecurityRecord;
     }
 }
